Consolidated Typescript code into functions for ease of readability and maintainability.
